chore(connectors): reorganize connectors
