+++ packages/connector-aliyun-dm/src/index.test.ts
@@ -1,8 +1,9 @@
 import { GetConnectorConfig } from '@logto/connector-types';
 
-import { AliyunDmConnector, AliyunDmConfig } from '.';
+import { AliyunDmConnector } from '.';
 import { mockedConfig } from './mock';
 import { singleSendMail } from './single-send-mail';
+import { AliyunDmConfig } from './types';
 
 const getConnectorConfig = jest.fn() as GetConnectorConfig<AliyunDmConfig>;
