+++ packages/connector-aliyun-dm/src/index.ts
@@ -9,31 +9,10 @@ import {
 } from '@logto/connector-types';
 import { assert } from '@silverhand/essentials';
 import { Response } from 'got';
-import { z } from 'zod';
 
 import { defaultMetadata } from './constant';
 import { singleSendMail } from './single-send-mail';
-import { SendEmailResponse } from './utils';
-
-/**
- * UsageType here is used to specify the use case of the template, can be either
- * 'Register', 'SignIn', 'ForgotPassword' or 'Test'.
- */
-const templateGuard = z.object({
-  usageType: z.string(),
-  subject: z.string(),
-  content: z.string(), // With variable {{code}}, support HTML
-});
-
-const configGuard = z.object({
-  accessKeyId: z.string(),
-  accessKeySecret: z.string(),
-  accountName: z.string(),
-  fromAlias: z.string().optional(),
-  templates: z.array(templateGuard),
-});
-
-export type AliyunDmConfig = z.infer<typeof configGuard>;
+import { SendEmailResponse, AliyunDmConfig, aliyunDmConfigGuard } from './types';
 
 export class AliyunDmConnector implements EmailConnector {
   public metadata: ConnectorMetadata = defaultMetadata;
@@ -45,7 +24,7 @@ export class AliyunDmConnector implements EmailConnector {
   }
 
   public validateConfig: ValidateConfig = async (config: unknown) => {
-    const result = configGuard.safeParse(config);
+    const result = aliyunDmConfigGuard.safeParse(config);
 
     if (!result.success) {
       throw new ConnectorError(ConnectorErrorCodes.InvalidConfig, result.error.message);
