+++ packages/connector-wechat-native/src/types.ts
@@ -1,0 +1,22 @@
+import { z } from 'zod';
+
+export const weChatNativeConfigGuard = z.object({ appId: z.string(), appSecret: z.string() });
+
+export type WeChatNativeConfig = z.infer<typeof weChatNativeConfigGuard>;
+
+export type AccessTokenResponse = {
+  access_token?: string;
+  openid?: string;
+  expires_in?: number; // In seconds
+  refresh_token?: string;
+  scope?: string;
+  errcode?: number;
+};
+
+export type UserInfoResponse = {
+  unionid?: string;
+  headimgurl?: string;
+  nickname?: string;
+  errcode?: number;
+  errmsg?: string;
+};
