+++ packages/connector-wechat/src/constant.ts
@@ -2,16 +2,11 @@ import path from 'path';
 
 import { ConnectorMetadata, ConnectorType, ConnectorPlatform } from '@logto/connector-types';
 import { getFileContents } from '@logto/shared';
-import { z } from 'zod';
 
 export const authorizationEndpoint = 'https://open.weixin.qq.com/connect/qrconnect';
 export const accessTokenEndpoint = 'https://api.weixin.qq.com/sns/oauth2/access_token';
 export const userInfoEndpoint = 'https://api.weixin.qq.com/sns/userinfo';
 export const scope = 'snsapi_login';
-
-export const weChatConfigGuard = z.object({ appId: z.string(), appSecret: z.string() });
-
-export type WeChatConfig = z.infer<typeof weChatConfigGuard>;
 
 // eslint-disable-next-line unicorn/prefer-module
 const currentPath = __dirname;
